Move config.ts to settings.ts
