+++ src/lib/base/Util/AppState.ts
@@ -1,4 +1,4 @@
-import { config, userSettingsLabels } from '../../../config/config';
+import { config, userSettingsLabels } from '../../../config/settings';
 import { createPersistedStore } from './PersistedStore';
 
 // The settingsStore is backed by values in config. Only values specified in userSettingsLabels
